+++ index.html
@@ -147,13 +147,9 @@
         </div>
       </div>
     </div>
-    <script src="./scripts/InitialCards.js"></script>
-    <script src="./scripts/Card.js"></script>
-    <script src="./scripts/CardList.js"></script>
-    <script src="./scripts/Popup.js"></script>
-    <script src="./scripts/UserInfo.js"></script>
-    <script src="./scripts/FormValidator.js"></script>
-    <script src="./scripts/Api.js"></script>
-    <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
+    <script
+      type="module"
+      src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"
+    ></script>
   </body>
 </html>
