+++ scripts/Card.js
@@ -1,4 +1,4 @@
-class Card {
+export default class Card {
   constructor(data) {
     this.data = data;
   }
