--- scripts/Review.md
@@ -1,65 +1,0 @@
-Добрый день. Работа опрятная, код читаемый, но функционал должным образом не работает. Также не все запросы
-к серверу имеют catch, что не очень хорошо (см. ниже)
-
-## Надо исправить:
-
-При открытии формы редактирования пользователя в самый первый раз, имя и занятия не подгружаются,
-там все еще старик Кусто.
-
-Какая-то беда с редактированием профиля, а все потому что вы не те параметры передаете.
-
-Протестируйте код в разных браузерах. Протестируйте при выключенном интернете, в консоль должны прилетать
-только те сообщения об ошибках, которые вы сами выводите, не должно быть ошибки скрипта из-за неопределенной
-переменной.
-
-Прежде чем вносить изменения в DOM обязательно убедитесь, что опреация на сервере прошла успешно,
-иначе пользователь будет думать что он, например, профиль обновил, но на самом деле вы ему только
-на старнице данные обновили.
-
-Допустим у вас есть метод Api:
-
-someApiMethod() {
-return fetch(...)
-.then(...)
-.catch(...)
-}
-
-catch тоже должен либо пробрасывать ошибку через throw или возвращать промис с ошибкой,
-потому что в месте вызова someApiMethod тоже нужно обработчик ошибок сделать. У Api свой есть,
-но фронтенд должен тоже на ошибки сервера адекватно реагировать когда методы Api вызывает:
-
-someApiMethod()
-.then(... что-то рисуем, меняем, закрываем)
-.catch(... ничего не меняем, не закрываем, сообщаем об ошибке)
-
-Вообще лююой fetch даже с GET запросом должен заканчиваться catch, ошибки бывают всегда. Их надо
-уметь обрабатывать максимально корректно для пользователя.
-
-Если обновление данных пользователя не сработало, попап не должен закрываться.
-
-Данные пользователя должны подгружаться на страницу, моковых данных быть не должно.
-
-Ну и в коде еще есть комментарии.
-
-## Можно лучше
-
-Большое количество параметов лучше передвать в метод или в конструктор через деструктуризацию.
-
-Например в коде:
-const newClass = new Class({ windowOne, userForm, popupObj })
-
-А внутри класса:
-constructor ({ userForm,popupObj, windowOne }) {...}
-
-И тогда порядок переменных будет неважен, это удобно
-
-Для слушателей событий пользуйтесь стрелочными функциями
-
-## Промежуточный итог
-
-Аккуратная работа, но кое-что не доделано + запутались с тем откуда юзеринфо брать при инициализации.
-Исправьте критические недочеты и присылайте, у вас это много времени не займет.
-
-## Вопрос
-
-Скажите пожалуйста, vs code настойчиво предлагает переделать промисы под async/await, это тот же синтаксический сахар, или как то скажется на функциональности?
